+++ xmake/actions/install/install.lua
@@ -36,7 +36,7 @@ function _install_binary(target)
     end
 
     -- the binary directory
-    local binarydir = path.join(_g.installdir, "bin")
+    local binarydir = path.join(target:installdir(), "bin")
 
     -- make the binary directory
     os.mkdir(binarydir)
@@ -54,10 +54,10 @@ function _install_library(target)
     end
 
     -- the library directory
-    local librarydir = path.join(_g.installdir, "lib")
+    local librarydir = path.join(target:installdir(), "lib")
 
     -- the include directory
-    local includedir = path.join(_g.installdir, "include")
+    local includedir = path.join(target:installdir(), "include")
 
     -- make the library directory
     os.mkdir(librarydir)
@@ -107,6 +107,15 @@ function _on_install_target(target)
     if target:get("enabled") == false then
         return 
     end
+
+    -- get install directory
+    local installdir = target:installdir()
+    if not installdir then
+        return 
+    end
+
+    -- trace
+    print("installing %s to %s ...", target:name(), installdir)
 
     -- build target with rules
     local done = false
@@ -200,13 +209,10 @@ function _install_target_and_deps(target)
 end
 
 -- install targets
-function main(targetname, installdir)
+function main(targetname)
 
     -- init finished states
     _g.finished = {}
-
-    -- init install directory
-    _g.installdir = installdir
 
     -- install the given target?
     if targetname and not targetname:startswith("__") then