+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `string.serialize` and `string.deserialize` to serialize and deserialize object, function and others.
 * Add `xmake g --menu`
+* [#283](https://github.com/tboox/xmake/issues/283): Add `target:installdir()` and `set_installdir()` api for target
 
 ### Changes
 
@@ -535,6 +536,7 @@
 
 * 添加`string.serialize`和`string.deserialize`去序列化，反序列化对象，函数以及其他类型
 * 添加`xmake g --menu`去图形化配置全局选项
+* [#283](https://github.com/tboox/xmake/issues/283): 添加`target:installdir()`和`set_installdir()`接口
 
 ### 改进